# openapi.yaml
openapi: 3.1.0
info:
  title: RFO Protocol API
  version: 1.0.0
  description: |
    ## RFO (Research For Optimization) Protocol
    
    A next-generation application-layer network protocol designed for AI web-scale agents.
    RFO enables structured, verifiable, and efficient content delivery between origin servers
    and AI agents through cryptographic site identification, quality scoring, and
    token-optimized payloads.
    
    ### Key Concepts
    - **Handshake**: Bidirectional protocol where the agent provides a domain URL and the engine
      compiles, verifies, and returns structured `.doc` or `.mdoc` payloads.
    - **Site ID**: Cryptographically derived, deterministic identity tied to domain + secret + time window.
    - **Quality Score**: Automated 0–100 scoring of content structure, verifiability, and AEO readiness.
    - **Payload Types**:
      - `.doc` — Full deep-knowledge layout (markdown, data tables, verification signature).
      - `.mdoc` — Token-optimized mini-document (< 1,500 tokens) for LLM context windows.
    - **Streaming**: Binary chunked transfer for large payloads with checksums.
    - **WebSocket**: Real-time pub/sub for domain update notifications.
    - **Capability Negotiation**: Protocol version and encoding negotiation (JSON / MessagePack).
    
    ### Authentication
    All protected endpoints require an `X-API-Key` header. Public endpoints (`/rfo/health`,
    `/rfo/capabilities`, `/rfo/negotiate`) do not require authentication.
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for client authentication. Obtain keys from your RFO deployment administrator.
        Format: `X-API-Key: <your-api-key>`
  schemas:
...
    RfoHeader:
      type: object
      required: [site_id, coordinates, quality_score]
      properties:
        site_id:
          type: string
          description: Cryptographically derived site identifier (SHA-256 hex, 64 chars)
          example: "a1b2c3d4e5f6..."
        coordinates:
          type: object
          additionalProperties:
            type: string
          description: Semantic coordinates extracted from content
          example: {"topic": "machine-learning", "subtopic": "transformers"}
        quality_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Automated content quality score (0-100)
          example: 85
    QaPair:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
          description: Extracted question from content
        answer:
          type: string
          description: Corresponding answer from content
    MiniDocPayload:
      type: object
      required: [summary, token_count, qa_pairs]
      description: Token-optimized mini-document for LLM context windows (< 1,500 tokens)
      properties:
        summary:
          type: string
          description: Concise summary of the page content
        token_count:
          type: integer
          description: Estimated token count
        qa_pairs:
          type: array
          items:
            $ref: '#/components/schemas/QaPair'
          description: Question-answer pairs for AEO
    FullDocPayload:
      type: object
      required: [raw_markdown, data_tables, verification_signature]
      description: Deep knowledge layout with markdown, data tables, and verification
      properties:
        raw_markdown:
          type: string
          description: Full content as structured markdown
        data_tables:
          type: array
          items:
            type: string
          description: Extracted data tables as markdown
        verification_signature:
          type: string
          description: HMAC signature for content verification
    PayloadType:
      type: string
      enum: [doc, mdoc]
      description: |
        - `doc` — Full deep-knowledge payload
        - `mdoc` — Token-optimized mini-document
    Payload:
      oneOf:
        - $ref: '#/components/schemas/FullDocPayload'
        - $ref: '#/components/schemas/MiniDocPayload'
      description: Either a FullDocPayload or MiniDocPayload

    # ── Handshake ─────────────────────────────────────────────────────────────
    HandshakeRequest:
      type: object
      required: [domain_url, coordinates, requested_payload, nonce, timestamp]
      properties:
        domain_url:
          type: string
          format: uri
          description: Target URL to compile
          example: "https://example.com"
        coordinates:
          type: object
          additionalProperties:
            type: string
          description: Client-provided semantic coordinates to merge
        requested_payload:
          $ref: '#/components/schemas/PayloadType'
        nonce:
          type: string
          description: Cryptographic nonce for replay protection (generated via /rfo/handshake)
          minLength: 32
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp when nonce was generated (must be within 5 minutes)
    HandshakeResponse:
      type: object
      required: [header, payload, processing_time_ms, nonce]
      properties:
        header:
          $ref: '#/components/schemas/RfoHeader'
        payload:
          $ref: '#/components/schemas/Payload'
        processing_time_ms:
          type: integer
          description: Server processing time in milliseconds
        nonce:
          type: string
          description: Echoed nonce from request
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
  headers:
    X-RFO-Protocol-Version:
      schema:
        type: string
      description: RFO protocol version used in the response
    X-RFO-Site-ID:
      schema:
        type: string
      description: Site ID of the compiled domain
paths:
...
  /rfo/handshake:
    post:
      tags: [Handshake]
      summary: Duplex handshake
      description: |
        Core protocol endpoint. The client provides a domain URL and the engine:
        1. Verifies nonce freshness (replay protection, 5-minute window)
        2. Checks cache (sub-5ms path)
        3. Fetches and parses the target URL
        4. Extracts semantic coordinates
        5. Generates cryptographic site ID
        6. Compiles .doc and .mdoc payloads
        7. Calculates quality score
        8. Caches the result
        9. Upserts the site record in the database
        10. Returns the compiled payload
      operationId: handshake
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandshakeRequest'
      responses:
        '200':
          description: Handshake successful
          headers:
            X-RFO-Protocol-Version:
              $ref: '#/components/headers/X-RFO-Protocol-Version'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandshakeResponse'
        '401':
          description: Invalid/expired nonce or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
        '502':
          description: Failed to fetch target URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /rfo/doc/{domain}:
    get:
      tags: [Payloads]
      summary: Get full .doc payload
      description: |
        Returns the compiled FullDocPayload for a domain. The domain must have
        been previously compiled via a handshake. Returns 404 if not found.
      operationId: getDoc
      security:
        - ApiKeyAuth: []
      parameters:
        - name: domain
          in: path
          required: true
          schema:
            type: string
          description: Domain URL (e.g., "https://example.com")
      responses:
        '200':
          description: Full document payload
          headers:
            X-RFO-Protocol-Version:
              $ref: '#/components/headers/X-RFO-Protocol-Version'
            X-RFO-Site-ID:
              $ref: '#/components/headers/X-RFO-Site-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullDocPayload'
        '404':
          description: Domain not compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - ApiKeyAuth: []